better handling of the __abouttofreeze recursion.
better handling of compilation after an error occurs

Example: a calls b calls c, b calls a.

a -> b -> c
^ ---

Previously if b had an error then c and a would fail to compile as well.
Now c will succeed -- it doesn't depend on anything that had an error in it.
However, a will fail since it depends on b for its behavior.