fixing a bug where includec would fail if C code used the same name for a type and a function. Include C now returns two arguments C,T = terralib.includec(...)
C will contain all the functions, and any types whose names do not conflict with the functions
T will contain all types.

Most of the time you can ignore the second argument.

Eventually we need to improve the handling of typedefs vs. struct definitions from C since there
are a few lurking problems (such as when struct Foo and Foo are not the same type).

fixes #24