fixing issue with name conflicts between typedefs Foo and struct Foo. includec now returns two _namespaces_. The first is the generic C namespace where functions, variables, and typedefs live. The second is the 'tagged' namespace where structures, unions, and enums live.
generic,tagged = terralib.includec()

like before, members of the tagged namespace that do not conflict with names in the
generic namespace will also be present in the generic type. So the only time
you need to use 'tagged' is when the name appears in both such as 'stat' on unix.