Added support for llvm/clang version 3.3.1
The official svn source for llvm and clang is actually version 3.3.1
when compiling against the release_33 branch. This commit adds support
for this point release when building terra.

The make file only converted the first period to an underscore. This has
been fixed so the llvm version is now reported as LLVM_3_3_1. The
appropriate preprocessor macros have been modified to allow this version
to follow the same path as LLVM_3_3.

All tests pass but I'm not sure if there are any sutble differences
between 3.3 and 3.3.1 that would make a difference to terra.