change path handling for terra to be more analogous to how Lua and C files are loaded
terralib.require is not removed, used 'require' instead.
package.terrapath is now the search path for Terra files (distinguished from package.path)
It is initialized to the enivronment variable TERRA_PATH or the default path './?.t' if not specified'
Use ';;' in TERRA_PATH to indicate that the default path should be appended.