Debian multarch compatibility.
Includes changes that mean the x86_64 compile will correctly run on a
32bit system with suitable library and kernel support.

If you have a multiarch aware llvm-dev and clang-dev the Makefile changes
will allow you to build Terra on a mostly 32bit host using:

    make TARGET_ARCH=-m64

The robustness of the Makefile has also been improved slightly so that
a compile that is started on a 32bit environment without the required
setup should fail early enough so it does not require a "make purge"
to successfully complete a 64bit compile.

To run the 64bit application on a 32bit user space you can use multiarch,
but you actually only need "gcc-multilib".

The 32bit packages needed to run terra can be installed using the
Debian/Ubuntu packages:

    apt-get install gcc-multilib lib64ncurses5 lib64z1

Note: Running 64bit applications on a 32bit system also requires that
the kernel be the "amd64" variant.