Upgrade LuaJIT to 2.1 (#426)
Upgrade LuaJIT to a version 2.1 from Git, currently pinned at commit 9143e86498436892cb4316550be4d45b68a61224. Also disable leaktest, as it depends on GC behavior that is not reliable.

If you want to use an official LuaJIT beta release, change the value of LUAJIT_COMMIT to an empty string and it will revert to using 2.1.0-beta3.