Move all Linux CI jobs to Docker (#600)
This avoids churn due to the deprecation schedule of GitHub Actions
runner images, and also fixes a number of issues that seem to result
from conflicts with the newer runner versions. The Docker build is now
fully featured and can build all configurations we use for the tests
(except for Make, which we relegate to the old build scripts). The
scripts are longer, but it's the same set of commands being executed
in the basic cases, so these are still minimal builds.